refactor: move wait_for_server() into client
